feat(worker): poll-and-dispatch — shared job-runner core + cloud run dispatch
Extract the connect/poll/execute core into execute/job-runner
(createApiConnection, fetchAndStoreSettings, buildMachineInfo,
appVersionCompatible, executeJob, runJob) so the poller and the executor
share one path.

buildJobHandler branches: with AP_CLOUD_RUN_URL set it forwards each polled
job to the Cloud Run executor (OIDC via google-auth, over safeHttp);
otherwise it runs jobs in a local sandbox pool as before. Zero behavior
change for the default LOCAL worker.
