+++ packages/server/worker/src/lib/worker.ts
@@ -1,40 +1,19 @@
 import { createServer } from 'http'
 import os from 'os'
-import { ActivepiecesError, isNil, spreadIfDefined, tryCatch } from '@activepieces/core-utils'
-import { Runtime, warmupPieces } from '@activepieces/sandbox-pool'
-import { apVersionUtil, onCallService, systemUsage, UNKNOWN_VERSION, wideEvent } from '@activepieces/server-utils'
-import { ConsumeJobRequest, createRpcClient, EngineResponseStatus, ExecutionMode, JobData, SandboxInformation, WebsocketServerEvent, WorkerMachineHealthcheckRequest, WorkerProps, WorkerSettingsResponse, WorkerToApiContract } from '@activepieces/shared'
-import { createLogger } from 'evlog'
+import { tryCatch } from '@activepieces/core-utils'
+import { Runtime, RuntimeExecutorInfo, warmupPieces } from '@activepieces/sandbox-pool'
+import { type ApLogger } from '@activepieces/server-utils'
+import { ConsumeJobRequest, WorkerToApiContract } from '@activepieces/shared'
 import { nanoid } from 'nanoid'
-import { io, Socket } from 'socket.io-client'
-import { getApiUrl, system, WorkerSystemProp } from './config/configs'
+import { Socket } from 'socket.io-client'
+import { system, WorkerSystemProp } from './config/configs'
 import { logger } from './config/logger'
-import { workerSettings } from './config/worker-settings'
-import { getHandler } from './execute/job-registry'
-import { JobContext, JobResult, JobResultKind } from './execute/types'
+import { createCloudRunDispatcher } from './execute/cloud-run-dispatch'
+import { appVersionCompatible, buildMachineInfo, createApiConnection, fetchAndStoreSettings, runJob } from './execute/job-runner'
 import { selectRuntime } from './runtime/runtime-factory'
 import { sandboxConfig } from './runtime/sandbox-config'
 
-
-const AP_VERSION = apVersionUtil.getCurrentRelease()
-
 const VERSION_MISMATCH_POLL_PAUSE_MS = 10_000
-
-let pagedForUnreadableWorkerVersion = false
-
-function pageOnceForUnreadableWorkerVersion(workerLog: typeof logger): void {
-    if (pagedForUnreadableWorkerVersion) {
-        return
-    }
-    pagedForUnreadableWorkerVersion = true
-    onCallService(workerLog, workerSettings.getSettings().PAGE_ONCALL_WEBHOOK).page({
-        code: 'WORKER_VERSION_READ_FAILED',
-        message: 'Worker could not read its release version from package.json (reported as 0.0.0); polling is paused and will NOT self-heal on reconnect until the deployment is fixed (check cwd/packaging)',
-        params: { workerVersion: AP_VERSION },
-    }).catch((pageError) => {
-        workerLog.error({ pageError }, 'Failed to send on-call page for unreadable worker version')
-    })
-}
 
 let socket: Socket | null = null
 let polling = false
@@ -48,21 +27,20 @@ let healthServerInstance: ReturnType<typeof createServer> | null = null
 
 let runtime: Runtime | null = null
 
+function getActiveExecutors(): RuntimeExecutorInfo[] {
+    return runtime?.getActiveExecutors() ?? []
+}
+
 export const worker = {
     async start({ apiUrl, socketUrl, workerToken, withHealthServer = false }: WorkerStartParams): Promise<void> {
         const workerGroupId = system.get(WorkerSystemProp.WORKER_GROUP_ID)
-        socket = io(socketUrl.url, {
-            auth: { token: workerToken, workerId, workerGroupId },
-            path: socketUrl.path,
-            transports: ['websocket'],
-            reconnection: true,
-        })
-
-        const apiClient = createRpcClient<WorkerToApiContract>(socket, 60_000)
+        const connection = createApiConnection({ socketUrl, workerToken, workerId, workerGroupId })
+        socket = connection.socket
+        const apiClient = connection.apiClient
 
         socket.on('connect', async () => {
             logger.info('Connected to API server via Socket.IO')
-            await fetchAndStoreSettings(socket!)
+            await fetchAndStoreSettings({ sock: socket!, workerId, workerHostname, getActiveExecutors })
             void warmupPiecesOnStartup(apiClient)
             void startPollingWorkers(apiClient).catch((err) => {
                 logger.error({ error: err }, 'Polling workers crashed unexpectedly')
@@ -116,39 +94,45 @@ async function startPollingWorkers(apiClient: WorkerToApiContract): Promise<void
     if (!Number.isInteger(rawConcurrency) || rawConcurrency < 1) {
         logger.warn({ rawConcurrency }, 'Invalid AP_WORKER_CONCURRENCY value, falling back to 1')
     }
-    runtime = selectRuntime({ concurrency, log: logger })
 
-    logger.info({ concurrency }, 'Starting polling workers')
+    const handleJob = buildJobHandler({ apiClient, concurrency })
 
-    const activeRuntime = runtime
     const workers = Array.from({ length: concurrency }, (_, workerIndex) =>
-        pollAndExecute(apiClient, activeRuntime, workerIndex, generation),
+        pollLoop({ apiClient, workerIndex, generation, handleJob }),
     )
     await Promise.all(workers)
 }
 
-async function pollAndExecute(apiClient: WorkerToApiContract, runtime: Runtime, workerIndex: number, generation: number): Promise<void> {
+// The worker either executes jobs in a local sandbox pool (LOCAL) or, when AP_CLOUD_RUN_URL is set,
+// forwards each polled job to a Cloud Run instance over HTTP. In dispatch mode the instance owns the
+// whole lifecycle (provision/run/progress/completeJob) over its own API connection, so the worker
+// builds no local runtime — it just polls and forwards. See ADR 0001.
+function buildJobHandler({ apiClient, concurrency }: BuildJobHandlerParams): HandleJob {
+    const cloudRunUrl = system.get(WorkerSystemProp.CLOUD_RUN_URL)
+    if (cloudRunUrl) {
+        const dispatcher = createCloudRunDispatcher({ baseUrl: cloudRunUrl })
+        logger.info({ concurrency, cloudRun: { url: cloudRunUrl } }, 'Starting workers in Cloud Run dispatch mode')
+        return ({ job, log }) => dispatcher.dispatch({ job, log })
+    }
+    runtime = selectRuntime({ concurrency, log: logger })
+    const activeRuntime = runtime
+    logger.info({ concurrency }, 'Starting workers in local execution mode')
+    return async ({ job, workerIndex, log }) => {
+        await runJob({ apiClient, runtime: activeRuntime, job, workerIndex, log })
+    }
+}
+
+async function pollLoop({ apiClient, workerIndex, generation, handleJob }: PollLoopParams): Promise<void> {
     const workerLog = logger.child({ workerIndex })
     workerLog.info('Polling worker started')
 
     while (polling && connectionGeneration === generation) {
-        const appVersion = workerSettings.getSettings().APP_VERSION
-        if (!apVersionUtil.versionsAreCompatible({ versionA: appVersion, versionB: AP_VERSION })) {
-            const versionUnreadable = appVersion === UNKNOWN_VERSION || AP_VERSION === UNKNOWN_VERSION
-            if (versionUnreadable) {
-                workerLog.error({ appVersion, workerVersion: AP_VERSION }, 'Pausing polling — a release version could not be read from package.json (reported as 0.0.0); this will NOT self-heal on reconnect, check the worker/app deployment (cwd/packaging)')
-            }
-            else {
-                workerLog.warn({ appVersion, workerVersion: AP_VERSION }, 'Connected app version mismatch — pausing polling until reconnect to a compatible app')
-            }
-            if (AP_VERSION === UNKNOWN_VERSION) {
-                pageOnceForUnreadableWorkerVersion(workerLog)
-            }
+        if (!appVersionCompatible({ log: workerLog })) {
             await sleep(VERSION_MISMATCH_POLL_PAUSE_MS)
             continue
         }
 
-        const { data: machineInfo, error: machineError } = await tryCatch(buildMachineInfo)
+        const { data: machineInfo, error: machineError } = await tryCatch(() => buildMachineInfo({ workerId, workerHostname, getActiveExecutors }))
         if (machineError) {
             workerLog.error({ error: machineError }, 'Failed to build machine info')
             await sleep(20000)
@@ -169,171 +153,8 @@ async function pollAndExecute(apiClient: WorkerToApiContract, runtime: Runtime, 
 
         workerLog.debug({ job: { id: job.jobId, type: job.jobData.jobType } }, 'Job received from poll')
 
-        const lockExtensionInterval = setInterval(() => {
-            void tryCatch(() => apiClient.extendLock({ jobId: job.jobId, token: job.token, queueName: job.queueName })).then(({ error }) => {
-                if (error) {
-                    workerLog.warn({ error, job: { id: job.jobId } }, 'Failed to extend lock')
-                }
-            })
-        }, 30_000)
-
-        const { data: result, error: execError } = await tryCatch(() =>
-            executeJob(apiClient, job, runtime, workerIndex),
-        )
-
-
-        const { error: completeError } = await tryCatch(() =>
-            apiClient.completeJob({
-                jobId: job.jobId,
-                token: job.token,
-                queueName: job.queueName,
-                status: execError
-                    ? EngineResponseStatus.INTERNAL_ERROR
-                    : result.status,
-                errorMessage: buildErrorMessage(execError ?? undefined, result ?? undefined),
-                logs: extractLogs(execError ?? undefined, result ?? undefined),
-                response: result?.kind === JobResultKind.SYNCHRONOUS ? result.response : undefined,
-            }),
-        )
-
-        clearInterval(lockExtensionInterval)
-
-        if (completeError) {
-            workerLog.error({ error: completeError, job: { id: job.jobId } }, 'Failed to complete job')
-        }
+        await handleJob({ job, workerIndex, log: workerLog })
     }
-}
-
-async function executeJob(apiClient: WorkerToApiContract, job: ConsumeJobRequest, runtime: Runtime, workerIndex: number): Promise<JobResult> {
-    const rawData = job.jobData
-    const jobData = JobData.parse(rawData)
-    const jobLogger = createLogger({
-        event: 'job.execute',
-        job: { id: job.jobId, type: jobData.jobType },
-        ...spreadIfDefined('requestId', 'requestId' in jobData ? jobData.requestId : 'httpRequestId' in jobData ? jobData.httpRequestId : undefined),
-        ...spreadIfDefined('project', 'projectId' in jobData && jobData.projectId != null ? { id: jobData.projectId } : undefined),
-        ...spreadIfDefined('platform', 'platformId' in jobData ? { id: jobData.platformId } : undefined),
-        ...spreadIfDefined('flow', 'flowId' in jobData ? { id: jobData.flowId } : undefined),
-        ...spreadIfDefined('flowRun', 'runId' in jobData ? { id: jobData.runId } : undefined),
-        ...spreadIfDefined('flowVersion', 'flowVersionId' in jobData ? { id: jobData.flowVersionId } : undefined),
-    })
-    return wideEvent.run({
-        logger: jobLogger,
-        fn: async () => {
-            const log = logger.child({ job: { id: job.jobId, type: jobData.jobType } })
-            const apiUrl = getApiUrl()
-            const { PUBLIC_URL: publicUrl } = await workerSettings.waitForSettings()
-            log.debug({ apiUrl, publicUrl }, 'Worker settings resolved')
-            const ctx: JobContext = {
-                apiClient,
-                runtime,
-                workerIndex,
-                jobId: job.jobId,
-                engineToken: job.engineToken,
-                internalApiUrl: apiUrl,
-                publicApiUrl: ensurePublicApiUrl(publicUrl),
-                log,
-            }
-            try {
-                const handler = getHandler(jobData.jobType)
-                log.debug({ handlerType: handler.jobType }, 'Executing job with handler')
-                const { data: result, error } = await tryCatch(() => handler.execute(ctx, jobData))
-                if (error) {
-                    log.error({ error }, 'Job execution failed')
-                    wideEvent.error(error)
-                    wideEvent.set({ outcome: 'failed' })
-                    throw error
-                }
-                log.debug('Job completed')
-                wideEvent.set({ outcome: 'success' })
-                return result
-            }
-            finally {
-                jobLogger.emit()
-            }
-        },
-    })
-}
-
-export function ensurePublicApiUrl(publicUrl: string): string {
-    if (publicUrl.endsWith('/api/')) return publicUrl
-    if (publicUrl.endsWith('/api')) return publicUrl + '/'
-    if (publicUrl.endsWith('/')) return publicUrl + 'api/'
-    return publicUrl + '/api/'
-}
-
-async function fetchAndStoreSettings(sock: Socket): Promise<void> {
-    const { data: request, error } = await tryCatch(buildMachineInfo)
-    if (error) {
-        logger.error({ error }, 'Failed to build machine info for settings fetch')
-        return
-    }
-    return new Promise<void>((resolve) => {
-        sock.emit(WebsocketServerEvent.FETCH_WORKER_SETTINGS, request, (response: WorkerSettingsResponse) => {
-            const localExecutionMode = system.get(WorkerSystemProp.EXECUTION_MODE)
-            if (!isNil(localExecutionMode)) {
-                response.EXECUTION_MODE = localExecutionMode
-            }
-            const workerGroupId = system.get(WorkerSystemProp.WORKER_GROUP_ID)
-            if (!isNil(workerGroupId)) {
-                const processSandboxedModes = [ExecutionMode.SANDBOX_PROCESS, ExecutionMode.SANDBOX_CODE_AND_PROCESS]
-                if (!processSandboxedModes.includes(response.EXECUTION_MODE as ExecutionMode)) {
-                    throw new Error(`Worker group "${workerGroupId}" requires AP_EXECUTION_MODE to be one of: ${processSandboxedModes.join(', ')}. Got: ${response.EXECUTION_MODE}`)
-                }
-                const reuseSandbox = system.get(WorkerSystemProp.REUSE_SANDBOX)
-                if (isNil(reuseSandbox)) {
-                    throw new Error(`Worker group "${workerGroupId}" requires AP_REUSE_SANDBOX to be set (true or false)`)
-                }
-            }
-            workerSettings.set(response)
-            logger.info({ environment: response.ENVIRONMENT, executionMode: response.EXECUTION_MODE }, 'Worker settings loaded')
-            resolve()
-        })
-    })
-}
-
-function getWorkerProps(): WorkerProps {
-    try {
-        const settings = workerSettings.getSettings()
-        return {
-            EXECUTION_MODE: settings.EXECUTION_MODE,
-            WORKER_CONCURRENCY: system.get(WorkerSystemProp.WORKER_CONCURRENCY)!,
-            SANDBOX_MEMORY_LIMIT: settings.SANDBOX_MEMORY_LIMIT,
-            REUSE_SANDBOX: system.get(WorkerSystemProp.REUSE_SANDBOX) ?? 'false',
-            version: AP_VERSION,
-        }
-    }
-    catch {
-        return {}
-    }
-}
-
-async function buildMachineInfo(): Promise<WorkerMachineHealthcheckRequest> {
-    const memInfo = await systemUsage.getContainerMemoryUsage()
-    const diskInfo = await systemUsage.getDiskInfo()
-    const cpuCores = await systemUsage.getCpuCores()
-    return {
-        workerId,
-        cpuUsagePercentage: systemUsage.getCpuUsage(),
-        diskInfo,
-        workerProps: getWorkerProps(),
-        ramUsagePercentage: memInfo.ramUsage,
-        totalAvailableRamInBytes: memInfo.totalRamInBytes,
-        totalCpuCores: cpuCores,
-        ip: workerHostname,
-        sandboxes: await buildSandboxInfo(),
-    }
-}
-
-async function buildSandboxInfo(): Promise<SandboxInformation[]> {
-    const activeExecutors = runtime?.getActiveExecutors() ?? []
-
-    return Promise.all(activeExecutors.map(async (executor) => ({
-        sandboxId: executor.sandboxId,
-        boxId: executor.boxId,
-        busy: executor.busy,
-        memoryUsageBytes: await systemUsage.getProcessTreeMemoryBytes(executor.pid),
-    })))
 }
 
 async function warmupPiecesOnStartup(apiClient: WorkerToApiContract): Promise<void> {
@@ -365,35 +186,9 @@ async function warmupPiecesOnStartup(apiClient: WorkerToApiContract): Promise<vo
     logger.info({ count: pieces.length }, 'Piece cache warmup complete')
 }
 
-function buildErrorMessage(execError: Error | undefined, result: JobResult | undefined): string | undefined {
-    if (execError) {
-        return execError.message
-    }
-    const isFailure = result?.kind === JobResultKind.SYNCHRONOUS && result.status !== EngineResponseStatus.OK
-    if (!isFailure) {
-        return undefined
-    }
-    return result.errorMessage
-}
-
-function extractLogs(execError: Error | undefined, result: JobResult | undefined): string | undefined {
-    if (execError instanceof ActivepiecesError) {
-        const params = execError.error.params as Record<string, unknown>
-        const parts: string[] = []
-        if (params?.['standardOutput']) parts.push(`stdout:\n${params['standardOutput']}`)
-        if (params?.['standardError']) parts.push(`stderr:\n${params['standardError']}`)
-        return parts.length > 0 ? parts.join('\n') : undefined
-    }
-    if (result && 'logs' in result) {
-        return result.logs
-    }
-    return undefined
-}
-
 function sleep(ms: number): Promise<void> {
     return new Promise((resolve) => setTimeout(resolve, ms))
 }
-
 
 function startHealthServer(): ReturnType<typeof createServer> {
     const port = Number(process.env[WorkerSystemProp.PORT] ?? system.get(WorkerSystemProp.PORT))
@@ -419,4 +214,18 @@ type WorkerStartParams = {
     socketUrl: { url: string, path: string }
     workerToken: string
     withHealthServer?: boolean
+}
+
+type HandleJob = (params: { job: ConsumeJobRequest, workerIndex: number, log: ApLogger }) => Promise<void>
+
+type BuildJobHandlerParams = {
+    apiClient: WorkerToApiContract
+    concurrency: number
+}
+
+type PollLoopParams = {
+    apiClient: WorkerToApiContract
+    workerIndex: number
+    generation: number
+    handleJob: HandleJob
 }
